+++ tools/migrate/__main__.py
@@ -18,7 +18,6 @@ import datetime
 import importlib.util
 import os
 import sys
-import textwrap
 import traceback
 from typing import Any, Dict, List, Optional
 
@@ -26,6 +25,8 @@ import boto3
 from mypy_boto3_dynamodb import DynamoDBClient, DynamoDBServiceResource
 from pynamodb.attributes import UnicodeAttribute, UTCDateTimeAttribute
 from pynamodb.models import Model
+
+from ._generator import create_migration_file
 
 # --- Config ---
 MIGRATIONS_DIR = os.path.join(os.getcwd(), "migrations")
@@ -46,6 +47,7 @@ class MigrationHistory(Model):
     applied_at = UTCDateTimeAttribute(null=False)
 
 
+# --- Migration history helpers ---
 def ensure_migration_history_table():
     if not MigrationHistory.exists():
         MigrationHistory.create_table(
@@ -74,7 +76,7 @@ def parse_revision_from_filename(fname: str) -> str:
 
 def import_migration_module(filepath: str):
     spec = importlib.util.spec_from_file_location("migration_module", filepath)
-    if spec is None:
+    if spec is None or spec.loader is None:
         raise ImportError(f"Could not load spec for {filepath}")
     mod = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(mod)  # type: ignore
@@ -102,7 +104,7 @@ def remove_revision_record(revision: str):
         pass
 
 
-# --- DynamoDB copy helper (for ALTER-like ops) ---
+# --- DynamoDB copy helper (optional) ---
 def copy_table(
     src_table_name: str,
     dst_table_name: str,
@@ -167,237 +169,6 @@ def copy_table(
 
     print("Table copy finished.")
     return True
-
-
-# --- File templates ---
-CREATE_TEMPLATE = """\
-# mypy: ignore-errors
-from pynamodb.models import Model
-from pynamodb.attributes import {attr_imports}  # type: ignore # noqa: F401
-
-revision = "{revision}"
-down_revision = {down_revision_repr}
-
-
-class {class_name}(Model):
-    class Meta:  # type: ignore
-        table_name = "{table_name}"
-        region = "{region}"
-
-    {hash_key_name} = {hash_key_type}(hash_key=True)  # 主キー属性
-
-    # 他の属性は add コマンドで追加してください
-
-
-def upgrade():
-    \"\"\"Apply: create table if not exists\"\"\"
-    if not {class_name}.exists():
-        {class_name}.create_table(read_capacity_units=5, write_capacity_units=5, wait=True)
-
-
-def downgrade():
-    \"\"\"Rollback: drop table (careful!)\"\"\"
-    if {class_name}.exists():
-        {class_name}.delete_table()
-"""
-
-
-def format_attribute_args(attributes: List[Dict[str, str]]) -> str:
-    # 例: id=user.id, email=user.email, ...
-    return ", ".join([f"{attr['name']}=user.{attr['name']}" for attr in attributes])
-
-
-ADD_ATTR_TEMPLATE = """
-# mypy: ignore-errors
-from pynamodb.models import Model
-from pynamodb.attributes import UnicodeAttribute, NumberAttribute, BooleanAttribute  # type: ignore # noqa: F401
-
-revision = "{revision}"
-down_revision = "{down_revision}"
-
-
-def upgrade():
-    \"\"\"Add attribute '{attr_name}' by creating a new table and copying data automatically.\"\"\"
-    # 旧テーブル
-    class Old{class_name}(Model):
-        class Meta:  # type: ignore
-            table_name = "{table_name}"
-            region = "{region}"
-    # Existing attributes
-{old_attributes}
-
-    # 新テーブル
-    class New{class_name}(Model):
-        class Meta:  # type: ignore
-            table_name = "{table_name}"
-            region = "{region}"
-    # Existing attributes plus new attribute
-{new_attributes}
-    # 新テーブル作成
-    if not New{class_name}.exists():
-        New{class_name}.create_table(billing_mode='PAY_PER_REQUEST', wait=True)
-
-    # 旧テーブルから新テーブルへデータコピー
-    for user in Old{class_name}.scan():
-        New{class_name}({attribute_args}, {attr_name}=getattr(user, '{attr_name}', None)).save()
-
-    # 旧テーブル削除
-    if Old{class_name}.exists():
-        Old{class_name}.delete_table()
-
-    print("Data copy to new table '{table_name}' completed. Old table has been automatically deleted.")
-
-
-def downgrade():
-    \"\"\"Remove attribute '{attr_name}' by copying data to a table without the attribute.\"\"\"
-    # 旧テーブル
-    class Old{class_name}(Model):
-        class Meta:  # type: ignore
-            table_name = "{table_name}"
-            region = "{region}"
-        # Existing attributes (including the attribute to be removed)
-{new_attributes}
-
-    # 新テーブル（属性なし）
-    class New{class_name}(Model):
-        class Meta:  # type: ignore
-            table_name = "{table_name}"
-            region = "{region}"
-        # Existing attributes (without the attribute to be removed)
-{old_attributes}
-
-    # 新テーブル作成
-    if not New{class_name}.exists():
-        New{class_name}.create_table(billing_mode='PAY_PER_REQUEST', wait=True)
-
-    # データコピー
-    for user in Old{class_name}.scan():
-        New{class_name}({attribute_args}).save()
-
-    # 旧テーブル削除
-    if Old{class_name}.exists():
-        Old{class_name}.delete_table()
-
-    print("Data copied to the original table '{table_name}'. The temporary table has been deleted automatically.")
-"""
-
-
-# --- Generator functions ---
-def format_attributes(
-    attributes: List[Dict[str, str]],
-    hash_key_name: str,
-    indent: int = 4,
-) -> str:
-    pad = " " * indent
-    lines = []
-    for attr in attributes:
-        if attr["name"] == hash_key_name:
-            lines.append(f"{pad}{attr['name']} = {attr['type']}(hash_key=True)")
-        else:
-            lines.append(f"{pad}{attr['name']} = {attr['type']}()")
-    return "\n".join(lines)
-
-
-def create_migration_file(
-    action: str,
-    table_name: str,
-    extra: Optional[str] = None,
-    down_rev: Optional[str] = None,
-    hash_key_name: str = "id",
-    hash_key_type: str = "UnicodeAttribute",
-    attr_imports: str = "UnicodeAttribute, NumberAttribute, BooleanAttribute, UTCDateTimeAttribute",
-    attributes: Optional[List[Dict[str, str]]] = None,
-):
-    """Generate a migration file from templates.
-
-    Before writing to disk, validate the generated Python source with
-    compile(..., mode='exec'). If a SyntaxError is detected, print a concise
-    error and abort without creating the file.
-    """
-    revision = now_revision_str()
-    fname = f"{revision}_{action}_{table_name}.py"
-    path = os.path.join(MIGRATIONS_DIR, fname)
-
-    down_revision_repr = f'"{down_rev}"' if down_rev else "None"
-    class_name = table_name.capitalize()
-    region = DEFAULT_REGION
-
-    if attributes is None:
-        attributes = [{"name": hash_key_name, "type": hash_key_type}]
-
-    if action == "create":
-        content = CREATE_TEMPLATE.format(
-            revision=revision,
-            down_revision_repr=down_revision_repr,
-            class_name=class_name,
-            table_name=table_name,
-            region=region,
-            hash_key_name=hash_key_name,
-            hash_key_type=hash_key_type,
-            attr_imports=attr_imports,
-        )
-    elif action in ("add", "remove", "update"):
-        if not extra:
-            raise ValueError("extra attribute specification required for add/remove")
-        attr_name, attr_type = extra.split(":", 1)
-        # Attribute declarations occur inside a class nested in a function; use 8-space indent
-        old_attributes = format_attributes(attributes, hash_key_name, indent=8)
-        new_attributes = format_attributes(
-            attributes + [{"name": attr_name, "type": attr_type}],
-            hash_key_name,
-            indent=8,
-        )
-        attribute_args = format_attribute_args(attributes)
-        content = ADD_ATTR_TEMPLATE.format(
-            revision=revision,
-            down_revision=down_rev or "",
-            class_name=class_name,
-            table_name=table_name,
-            region=region,
-            attr_name=attr_name,
-            attr_type=attr_type,
-            old_attributes=old_attributes,
-            new_attributes=new_attributes,
-            attribute_args=attribute_args,
-        )
-    else:
-        # generic template
-        content = textwrap.dedent(
-            f"""\
-        revision = "{revision}"
-        down_revision = {down_revision_repr}
-
-
-        def upgrade():
-            pass
-
-
-        def downgrade():
-            pass
-        """
-        )
-    # Validate syntax before writing
-    try:
-        # Use the final filename for clearer error messages
-        compile(content, fname, "exec")
-    except SyntaxError as e:
-        # Do not write the file; surface a clear message and fail
-        err_line = e.text.strip() if e.text else ""
-        print(
-            "SyntaxError in generated migration:\n"
-            f"  file: {fname}\n"
-            f"  line: {e.lineno}, column: {e.offset}\n"
-            f"  msg: {e.msg}\n"
-            f"  code: {err_line}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    # Write only if syntax is valid
-    with open(path, "w", encoding="utf-8") as f:
-        f.write(content)
-    print(f"Generated migration: {path}")
-    return path
 
 
 # --- Runner logic ---
